+++ src/sensors/MeterTeros11.h
@@ -38,6 +38,12 @@
  * power is connected to the _white_ cable, data to _red_, and ground to the
  * unshielded cable.
  *
+ * @warning Coming from the factory, METER sensors are set at SDI-12 address
+ * '0'.  They also output a "DDI" serial protocol string on each power up.
+ * This library *disables the DDI output string* on all newer METER sensors
+ * that support disabling it.  After using a METER sensor with ModularSensors,
+ * you will need to manually re-enable the DDI output if you wish to use it.
+ *
  * @section sensor_teros11_datasheet Sensor Datasheet
  * Documentation for the SDI-12 Protocol commands and responses for the Meter
  * Teros 11 can be found at:
@@ -100,6 +106,9 @@
 /// @brief Sensor::_measurementTime_ms; the Teros 11 takes25 ms to 50 ms to
 /// complete a measurement.
 #define TEROS11_MEASUREMENT_TIME_MS 50
+/// @brief Extra wake time required for an SDI-12 sensor between the "break"
+/// and the time the command is sent.  The Terros-11 requires no extra time.
+#define TEROS11_EXTRA_WAKE_TIME_MS 0
 /**@}*/
 
 /**
@@ -247,7 +256,8 @@ class MeterTeros11 : public SDI12Sensors {
         : SDI12Sensors(SDI12address, powerPin, dataPin, measurementsToAverage,
                        "MeterTeros11", TEROS11_NUM_VARIABLES,
                        TEROS11_WARM_UP_TIME_MS, TEROS11_STABILIZATION_TIME_MS,
-                       TEROS11_MEASUREMENT_TIME_MS) {}
+                       TEROS11_MEASUREMENT_TIME_MS,
+                       TEROS11_EXTRA_WAKE_TIME_MS) {}
     /**
      * @copydoc MeterTeros11::MeterTeros11
      */
@@ -256,7 +266,8 @@ class MeterTeros11 : public SDI12Sensors {
         : SDI12Sensors(SDI12address, powerPin, dataPin, measurementsToAverage,
                        "MeterTeros11", TEROS11_NUM_VARIABLES,
                        TEROS11_WARM_UP_TIME_MS, TEROS11_STABILIZATION_TIME_MS,
-                       TEROS11_MEASUREMENT_TIME_MS) {}
+                       TEROS11_MEASUREMENT_TIME_MS,
+                       TEROS11_EXTRA_WAKE_TIME_MS) {}
     /**
      * @copydoc MeterTeros11::MeterTeros11
      */
@@ -265,7 +276,8 @@ class MeterTeros11 : public SDI12Sensors {
         : SDI12Sensors(SDI12address, powerPin, dataPin, measurementsToAverage,
                        "MeterTeros11", TEROS11_NUM_VARIABLES,
                        TEROS11_WARM_UP_TIME_MS, TEROS11_STABILIZATION_TIME_MS,
-                       TEROS11_MEASUREMENT_TIME_MS) {}
+                       TEROS11_MEASUREMENT_TIME_MS,
+                       TEROS11_EXTRA_WAKE_TIME_MS) {}
     /**
      * @brief Destroy the Meter Teros 11 object
      */