+++ ChangeLog.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ****
 
+## v0.28.4 (2021-05-05) [SDI-12 Timing Sensor Customization](https://github.com/EnviroDIY/ModularSensors/releases/tag/v0.28.4)
+
+### Downloads
+
+**[ModularSensors_Dependencies_v0.28.4.zip](https://github.com/EnviroDIY/ModularSensors/releases/download/v0.28.4/ModularSensors_Dependencies_v0.28.4.zip)**
+- contains this version of ModularSensors _and the proper versions of all of its dependencies_.
+
+### Improvements
+- Allow each SDI-12 sensor to set the necessary command delay for that sensor.
+  - Per protocol, sensors are allowed to take up to 100ms after receiving a break before being ready to receive a command.
+This allows each sensor to specify what delay it needs.
+This was added to support conflicting delay needs; the RDO needed a short delay, the newest Meter sensors do not respond properly if the delay is added.
+- For SDI-12 sensors, add repeated attempts to start a measurement if the first attempt unexpectedly fails.
+
 ## v0.28.3 (2021-03-24) [Valid version number](https://github.com/EnviroDIY/ModularSensors/releases/tag/v0.28.3)
 
 ### Downloads